# server/app/swagger_spec.json
{
  "swagger": "2.0",
  "info": {
    "title": "Streamix API",
    "description": "API para plataforma de videos tipo YouTube",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT. Ejemplo: Authorization: Bearer <token>"
    }
  },
  "security": [{"Bearer": []}],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener todos los usuarios",
        "responses": {"200": {"description": "Lista de usuarios"}}
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear usuario (admin)",
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": {
            "type": "object",
            "required": ["username", "email", "password"],
            "properties": {
              "username": {"type": "string"},
              "email": {"type": "string"},
              "password": {"type": "string"}
            }
          }
        }],
        "responses": {"201": {"description": "Usuario creado"}}
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener usuario por ID",
        "parameters": [{"name": "user_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Usuario encontrado"}, "404": {"description": "Usuario no encontrado"}}
      },
      "put": {
        "tags": ["Usuarios"],
        "summary": "Actualizar usuario",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "user_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Usuario actualizado"}, "403": {"description": "No autorizado"}, "404": {"description": "Usuario no encontrado"}}
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "user_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Usuario eliminado"}, "403": {"description": "No autorizado"}}
      }
    },
    "/videos": {
      "get": {
        "tags": ["Videos"],
        "summary": "Obtener todos los videos",
        "parameters": [
          {"name": "page", "in": "query", "type": "integer", "default": 1},
          {"name": "per_page", "in": "query", "type": "integer", "default": 10}
        ],
        "responses": {"200": {"description": "Lista de videos"}}
      },
      "post": {
        "tags": ["Videos"],
        "summary": "Subir video con miniatura",
        "description": "Sube un nuevo video con su archivo de video y miniatura. Requiere autenticación.",
        "security": [{"Bearer": []}],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "formData",
            "name": "title",
            "type": "string",
            "required": true,
            "description": "Título del video"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "Descripción del video (opcional)"
          },
          {
            "in": "formData",
            "name": "video",
            "type": "file",
            "required": true,
            "description": "Archivo de video (mp4, mkv, webm, avi)"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file",
            "required": true,
            "description": "Archivo de miniatura (png, jpg, jpeg, webp)"
          }
        ],
        "responses": {
          "201": {
            "description": "Video creado exitosamente",
            "schema": {
              "type": "object",
              "properties": {
                "message": {"type": "string", "example": "Video created"},
                "video": {"type": "object"}
              }
            }
          },
          "400": {"description": "Datos inválidos, archivos faltantes o formatos no permitidos"},
          "403": {"description": "No autorizado"},
          "500": {"description": "Error del servidor"}
        }
      }
    },
    "/videos/{video_id}": {
      "get": {
        "tags": ["Videos"],
        "summary": "Obtener video por ID",
        "parameters": [{"name": "video_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Video encontrado"}, "404": {"description": "Video no encontrado"}}
      },
      "put": {
        "tags": ["Videos"],
        "summary": "Actualizar video",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "video_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Video actualizado"}, "403": {"description": "No autorizado"}}
      },
      "delete": {
        "tags": ["Videos"],
        "summary": "Eliminar video",
        "security": [{"Bearer": []}],
        "parameters": [{"name": "video_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Video eliminado"}, "403": {"description": "No autorizado"}}
      }
    },
    "/users/{user_id}/videos": {
      "get": {
        "tags": ["Videos"],
        "summary": "Obtener videos de usuario",
        "parameters": [{"name": "user_id", "in": "path", "required": true, "type": "integer"}],
        "responses": {"200": {"description": "Videos del usuario"}}
      }
    },
...
    "/users/{user_id}/avatar": {
      "post": {
        "tags": ["Usuarios"],
        "summary": "Subir avatar de usuario",
        "description": "Sube una imagen de perfil (avatar) para el usuario. Solo el propio usuario puede subir su avatar.",
        "security": [{"Bearer": []}],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del usuario"
          },
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "required": true,
            "description": "Archivo de imagen (png, jpg, jpeg, gif, webp)"
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar subido exitosamente",
            "schema": {
              "type": "object",
              "properties": {
                "message": {"type": "string", "example": "Avatar uploaded successfully"},
                "avatar_url": {"type": "string", "example": "http://localhost:5000/public/profiles/1/avatar.jpg"},
                "usuario": {"type": "object"}
              }
            }
          },
          "400": {"description": "Archivo inválido o tipo no permitido"},
          "403": {"description": "No autorizado"},
          "500": {"description": "Error del servidor"}
        }
      }
    },
...
    "/users/me/subscriptions/videos": {
      "get": {
        "tags": ["Listas"],
        "summary": "Obtener videos de suscripciones",
        "security": [{"Bearer": []}],
        "responses": {"200": {"description": "Videos de canales suscritos"}}
      }
    }
  }
}